+++ lilya/cli/cli.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import typing
+from pathlib import Path
 
 import click
 from sayer import Sayer
@@ -19,6 +20,7 @@ from lilya.cli.directives.operations.runserver import runserver as runserver  # 
 from lilya.cli.directives.operations.shell import shell as shell  # noqa
 from lilya.cli.directives.operations.show_urls import show_urls as show_urls  # noqa
 from lilya.cli.groups import DirectiveGroup
+from lilya.cli.utils import get_custom_directives_to_cli
 
 help_text = """
 Lilya command line tool allowing to run Lilya native directives or
@@ -72,3 +74,15 @@ lilya_cli.add_command(create_app)
 lilya_cli.add_command(create_deployment)
 lilya_cli.add_command(shell)
 lilya_cli.add_app("mail", mail)
+
+
+# Load custom directives if any
+application_directives = get_custom_directives_to_cli(str(Path.cwd()))
+
+# Add application directives to the CLI
+if application_directives:
+    for _, command in application_directives.items():
+        if isinstance(command, Sayer):
+            lilya_cli.add_app(command._group.name, command)
+        else:
+            lilya_cli.add_command(command)